backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-middlewares==2.4.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudflared==1.0.0.2
cryptography==44.0.2
fastapi==0.115.12
firebase==4.0.1
firebase-admin==6.7.0
Flask==3.1.0
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.4.3
packaging==24.2
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.1.0
pytube==15.0.0
redis==5.2.1
requests==2.32.3
rsa==4.9
setuptools==78.1.0
setuptools-scm==8.2.0
sniffio==1.3.1
spotipy==2.25.1
starlette==0.46.2
Telethon==1.39.0
typing-inspection==0.4.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
yarl==1.19.0
youtube-search==2.1.2
yt-dlp==2025.3.31
